LATX, fix: Flush SMC reload tree with TB reset
Background:
An asynchronous tb_flush() cleared the SMC reload tree before its callback could establish that the captured flush count was still current. Nodes inserted after queuing could therefore outlive an accepted flush.

Solution:
Extract the TB flush coordinator so its asynchronous path can be tested directly. Clear the SMC reload tree only after do_tb_flush() accepts the captured count, while holding mmap_lock() and before the TB and TCG region resets. Stale callbacks leave the tree untouched.

RED:
test-latx-tb-flush-smc-reload-async exited 134 on the unfixed ordering:
ERROR: assertion failed (smc_reload_tree_get_node_count() == 0): (1 == 0)

GREEN:
test-latx-tb-flush-smc-reload-async exits 0. The focused LATX SMC suite passes, and libqemu-x86_64-linux-user.fa.p/accel_tcg_tb-flush.c.o builds.
