+++ accel/tcg/internal.h
@@ -20,4 +20,6 @@ void QEMU_NORETURN cpu_io_recompile(CPUState *cpu, uintptr_t retaddr);
 TranslationBlock *tb_link_page(TranslationBlock *tb, tb_page_addr_t phys_pc,
     tb_page_addr_t phys_page2);
 
+void tb_flush_remove_all(void);
+
 #endif /* ACCEL_TCG_INTERNAL_H */
